requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
duckdb>=1.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
typer>=0.12.0
rich>=13.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
supabase>=2.0.0
pyjwt>=2.8.0
langgraph>=1.0.0
langgraph-checkpoint>=3.0.0
langgraph-checkpoint-sqlite>=3.0.0
